Add seperate page for the java support
Let's have individual pages for the different languages
so that we can elaborate more on an individual
langauge support in parca-agent.

This commit adds one such page for the java support. In the
future we will have seperate pages for other langauges
so that users don't have figure it out by looking into
demo Dockerfiles.

Signed-off-by: Vaishali Thakkar <me.vaishalithakkar@gmail.com>